Move Checkbox close to SaveForm, which uses it
